perf ui/hist: Fix output for 100.00%
Current hpp format functions assume that the output will fit to 6
character including % sign (XX.YY%) so used "%5.2f%%" as a format
string.  However it might be the case if collapsing resulted in a
single entry which has 100.00% (7 character) of period. In this case
the output will be shifted by 1 character.

Signed-off-by: Namhyung Kim <namhyung@kernel.org>