perf tools: Convert to LIBPYTHON_SUPPORT
For building perf without python scripting, we can set NO_LIBPYTHON=1
as a argument of make.  It then defines NO_LIBPYTHON macro for C code
to do the proper handling.  However it usually used in a negative
semantics - e.g. #ifndef - so we saw double negations which can be
misleading.  Convert it to a positive form to make it more readable.
Also add _SUPPORT suffix for consistency.

Signed-off-by: Namhyung Kim <namhyung@kernel.org>