perf tools: Convert to DEMANGLE_SUPPORT
For building perf without name demangling, we can set NO_DEMANGLE=1 as
a argument of make.  It then defines NO_DEMANGLE macro for C code to
do the proper handling.  However it usually used in a negative
semantics - e.g. #ifndef - so we saw double negations which can be
misleading.  Convert it to a positive form to make it more readable.

Also add _SUPPORT suffix for consistency and make PACKAGE variable has
a string value "perf".

Cc: Markus Trippelsdorf <markus@trippelsdorf.de>
Cc: Mike Frysinger <vapier@gentoo.org>
Signed-off-by: Namhyung Kim <namhyung@kernel.org>