perf tools: Convert to LIBPERL_SUPPORT
For building perf without perf scripting, we can set NO_LIBPERL=1 as a
argument of make.  It then defines NO_LIBPERL macro for C code to do
the proper handling.  However it usually used in a negative semantics
- e.g. #ifndef - so we saw double negations which can be misleading.
Convert it to a positive form to make it more readable.

Also add _SUPPORT suffix for consistency and fix a few of indentation
in Makefile.

Signed-off-by: Namhyung Kim <namhyung@kernel.org>