perf gtk/annotate: Support multiple event annotation
Show multiple annotation result for each evsel.  Each result
represents the most frquently sampled symbol/function for the evsel
and it will be shown in a tab window.

For this add a reference to main container (notebook) to the pgctx.
At the first call to annotate browser, hist_entry__find_annotations()
will setup a new browser, and next calls will add new tabs to the
browser.  But it requires final perf_gtk__show_annotations() to start
processing GUI events.

Cc: Borislav Petkov <bp@alien8.de>
Cc: Pekka Enberg <penberg@kernel.org>
Signed-off-by: Namhyung Kim <namhyung@kernel.org>