perf sort: Separate out branch stack specific sort keys
Current perf report gets segmentation fault when a branch stack
specific sort key is provided by --sort option to a perf.data file
which contains no branch infomation.  It's because those sort keys
reference branch info of a hist entry unconditionally.  Maybe we can
change it checks whether such branch info is valid or not.  But if the
branch stacks are not recorded, it'd be nop.  Thus it'd be better to
make those keys are unselectable.

This patch separates those keys to a different dimension array, so
that if user passes such a key to a file which has no branch stack
will get following message rather than a segfault.

  Error: Invalid --sort key: `symbol_from'

Reported-by: Stefan Beller <stefanbeller@googlemail.com>
Cc: Stephane Eranian <eranian@google.com>
Signed-off-by: Namhyung Kim <namhyung@kernel.org>